Refactor MomentsHelper and create MomentsFormHelper
